Metodo para multar

+++ TallerDePOO/src/main/java/Biblioteca/Docente.java
@@ -54,7 +54,7 @@ public class Docente extends Lector {
 
 	public Docente(String nom, String apellido, String tipDoc, int numDoc, String correo, int numCel, LocalDate fecNac,
 			String sex, String nac, String domi, int codPos, String dep, String localidad, int canMulta,
-			Boolean estaMultado, int diasMultado, String carrera) {
+			Boolean estaMultado, Long diasMultado, String carrera) {
 		super(nom, apellido, tipDoc, numDoc, correo, numCel, fecNac, sex, nac, domi, codPos, dep, localidad, canMulta,
 				estaMultado, diasMultado);
 		this.carrera = carrera;
